migrate ConsumerProResponseNotification
